refact(userdata): change user data storing format, now match .json format
feat: add System destructor and delete call, preventing undefined behaviours

fix(files.cpp): fix meta file reading error

reading meta once for loading metadata was enabling eofbit and failbit and failing further file pointer seeks and reads

+++ core/include/files.h
@@ -48,11 +48,11 @@ class FileHandles {
 
 		// data operations
 
-		void storeMetadata(const CharBuffer&, int, int);
-		int readMetadata(CharBuffer&, int);
-		int getOffset(int);
-		int storeCredentials(const SecureCharBuffer&, const CharBuffer&, const SecureCharBuffer&, const CharBuffer&);
-		bool retrieveCredentials(SecureCharBuffer&, CharBuffer&, SecureCharBuffer&, CharBuffer&, int);
+		void storeMetadata(const CharBuffer&, uint64_t, int);
+		uint64_t readMetadata(CharBuffer&, int);
+		uint64_t getOffset(int);
+		uint64_t storeCredentials(const SecureCharBuffer&, const CharBuffer&, const SecureCharBuffer&, const CharBuffer&);
+		bool retrieveCredentials(SecureCharBuffer&, CharBuffer&, SecureCharBuffer&, CharBuffer&, uint64_t);
 };
 
 #endif // !FILES_H
